requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.1.post1
streamlit==1.33.0
streamlit-authenticator==0.2.3
PyYAML==6.0.1
